Split out ratings views & tests.
